patina_adv_logger: Refactor core logic from component logic
Currently the adv logger crate exposes all of the component and integration
test, etc. by default. However, for consumers wanting to log to the advanced
from more minimal environments, this is not ideal. This commit refactors
this base support to be the minimal core logic, and then a component feature
on top used by DXE.

Base: Exposes the core logger logic, and nothing else. Does no use alloc.

Component: Exposes the component, protocol, and integrations test logic.

To achieve this, this refactors the write and reader into separate modules
and simplifies to use a read-only and write-only paradigm to simplify the
wrapper logic.

With this change, consumers can use patina_adv_logger without default features
to get a minimal advanced logger implementation.

cleanup pub use
