+++ components/patina_adv_logger/src/lib.rs
@@ -7,14 +7,23 @@
 #![cfg_attr(all(not(feature = "std"), not(doc)), no_std)]
 #![feature(coverage_attribute)]
 
+#[cfg(any(feature = "alloc", test, doc))]
 extern crate alloc;
 
-pub mod component;
+mod memory_log;
+mod writer;
+
 pub mod logger;
+
+#[cfg(any(doc, feature = "reader"))]
+pub mod reader;
+
+#[cfg(any(doc, feature = "component"))]
+pub mod component;
+#[cfg(feature = "component")]
+mod integration_test;
+#[cfg(any(doc, feature = "component"))]
 pub mod protocol;
 
 #[cfg(any(doc, feature = "std"))]
 pub mod parser;
-
-mod integration_test;
-mod memory_log;
